+++ apps/directory/src/components/report-components/reportInterfaces.ts
@@ -7,3 +7,32 @@ export interface IContactInformation {
   address: string;
   country: { name: string; label: string };
 }
+
+export interface IReportDetail {
+  value: string;
+  type:
+    | "string"
+    | "email"
+    | "url"
+    | "bool"
+    | "list"
+    | "phone"
+    | "report"
+    | "string-with-key";
+  label?: string;
+  badgeColor?: IBadgeColor; // is this actually being used or dead code?
+}
+
+interface IBadgeColor {
+  type: EBadgeColors;
+}
+
+enum EBadgeColors {
+  "success",
+  "warning",
+  "info",
+  "secondary",
+  "danger",
+  "light",
+  "dark",
+}
